+++ packages/gitlab/src/components/GitlabCI/GitlabCI.tsx
@@ -8,20 +8,24 @@ import {
     MergeRequestsTable,
     MergeRequestStats,
     PipelinesTable,
+    ReleasesCard,
 } from '../widgets';
 
 export const GitlabCI = () => (
     <Page themeId="tool">
         <Content>
             <Grid container spacing={6} direction="row" alignItems="stretch">
-                <Grid item sm={12} md={6} lg={4}>
+                <Grid item sm={12} md={3} lg={3}>
                     <PeopleCard />
                 </Grid>
-                <Grid item sm={12} md={6} lg={4}>
+                <Grid item sm={12} md={3} lg={3}>
                     <LanguagesCard />
                 </Grid>
-                <Grid item sm={12} md={6} lg={4}>
+                <Grid item sm={12} md={3} lg={3}>
                     <MergeRequestStats />
+                </Grid>
+                <Grid item sm={12} md={3} lg={3}>
+                    <ReleasesCard />
                 </Grid>
                 <Grid item md={12}>
                     <PipelinesTable />